@@ -64,25 +64,28 @@
     "express": "^4.16.2",
     "history": "^4.7.2",
     "http-status-enum": "^1.0.2",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-redux": "^5.0.6",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
+    "react-redux": "^5.0.7",
     "react-router-dom": "^4.2.2",
     "react-router-redux": "^5.0.0-alpha.9",
-    "redux": "^3.7.2",
-    "redux-observable": "^0.17.0",
-    "rxjs": "^5.5.6"
+    "redux": "^4.0.0",
+    "redux-observable": "^0.18.0",
+    "rxjs": "^5.5.6",
+    "styled-components": "^3.2.6"
   },
   "devDependencies": {
+    "@types/enzyme": "^3.1.10",
     "@types/express": "^4.11.0",
     "@types/history": "^4.6.2",
     "@types/jest": "^22.0.1",
     "@types/nock": "^9.1.1",
-    "@types/react": "^16.0.34",
-    "@types/react-dom": "^16.0.3",
-    "@types/react-hot-loader": "^3.0.5",
-    "@types/react-redux": "^5.0.14",
-    "@types/react-router-dom": "^4.2.3",
+    "@types/react": "^16.3.13",
+    "@types/react-dom": "^16.0.5",
+    "@types/react-hot-loader": "^4.1.0",
+    "@types/react-redux": "^5.0.19",
+    "@types/react-router-dom": "^4.2.6",
+    "@types/react-test-renderer": "^16.0.1",
     "awesome-typescript-loader": "^4.0.0",
     "babel": "^6.23.0",
     "babel-cli": "^6.26.0",
@@ -104,24 +105,24 @@
     "gitbook-plugin-github": "^3.0.0",
     "gitbook-plugin-prism": "^2.3.0",
     "gitbook-plugin-theme-default": "^1.0.7",
-    "husky": "^0.14.3",
+    "husky": "^1.0.0-rc.2",
     "jest": "^22.0.5",
     "lint-staged": "^7.0.0",
     "nock": "^9.1.6",
-    "node-sass": "^4.7.2",
     "prettier": "^1.10.2",
     "react-hot-loader": "^4.0.0",
-    "react-test-renderer": "^16.2.0",
-    "sass-lint": "^1.12.1",
-    "sass-loader": "^6.0.6",
+    "react-test-renderer": "^16.3.2",
     "source-map-loader": "^0.2.3",
-    "style-loader": "^0.20.0",
+    "stylelint": "^9.2.0",
+    "stylelint-config-recommended": "^2.1.0",
+    "stylelint-config-styled-components": "^0.1.1",
+    "stylelint-processor-styled-components": "^1.3.1",
     "ts-jest": "^22.0.1",
     "tslint": "^5.8.0",
     "tslint-config-prettier": "^1.6.0",
     "tslint-react": "^3.3.3",
     "typescript": "^2.6.2",
-    "webpack": "^4.0.0",
+    "webpack": "^4.7.0",
     "webpack-dev-server": "^3.0.0"
   }
 }
