@@ -1,7 +1,8 @@
-// Import adapter for enzyme
-var enzyme = require('enzyme');
-var Adapter = require('enzyme-adapter-react-16');
-enzyme.configure({ adapter: new Adapter() })
+// Enzyme adaptations
+import * as Enzyme from 'enzyme'
+import * as Adapter from 'enzyme-adapter-react-16'
+
+Enzyme.configure({ adapter: new Adapter() });
 
 // Log all jsDomErrors when using jsdom testEnvironment
 window._virtualConsole && window._virtualConsole.on('jsdomError', function (error) {
