@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-const Loader: React.StatelessComponent<undefined> = () =>
+const Loader: React.StatelessComponent = () =>
     <div className="loader">
         <div className="loader__spinner" />
     </div>;
