@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-const PageNotFound: React.StatelessComponent<undefined> = () =>
+const PageNotFound: React.StatelessComponent = () =>
     <div className="page-not-found">404 - page not found</div>;
 
 export default PageNotFound;
