@@ -1,18 +1,18 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
 import createHistory from 'history/createBrowserHistory';
-import AppView from '../AppView';
+import { App } from '../App';
 
-describe('AppView', () => {
+describe('App', () => {
     const index = (
-        <AppView
+        <App
             match={{ params: undefined, isExact: true, path: '', url: '' }}
             location={{ pathname: '', search: '', state: {}, hash: '', key: '' }}
             history={createHistory()}
         />
     );
     const notFound = (
-        <AppView
+        <App
             match={{ params: undefined, isExact: false, path: '', url: '' }}
             location={{ pathname: '', search: '', state: {}, hash: '', key: '' }}
             history={createHistory()}
