@@ -1,25 +1,29 @@
 import * as React from 'react';
 import { RouteComponentProps } from 'react-router-dom';
+import { connect } from 'react-redux';
+
 import Todo from '../../common/Todo';
 import TodoComponent from '../../components/TodoComponent';
 import Button from '../../components/Button';
 import Loader from '../../components/Loader';
+import { State } from '../../redux/reducer';
+import { setTitle, saveTodo, setDone } from './IndexReducer';
 
-export interface IIndexState {
+export interface IndexState {
     title: string;
     todos: Todo[];
     loading: boolean;
 }
 
-export interface IIndexDispatch {
+export interface IndexDispatch {
     setTitle(n: string): void;
     saveTodo(): void;
     setDone(i: number): void;
 }
 
-export type IIndexProps = IIndexState & IIndexDispatch & RouteComponentProps<undefined>;
+export type IndexProps = IndexState & IndexDispatch & RouteComponentProps<undefined>;
 
-const IndexView: React.StatelessComponent<IIndexProps> = ({ title, todos, loading, setTitle, saveTodo, setDone }) =>
+export const Index: React.StatelessComponent<IndexProps> = ({ title, todos, loading, setTitle, saveTodo, setDone }) =>
     <main className="index">
         {loading && <Loader />}
         <h1 className="index__header">Todo app</h1>
@@ -43,4 +47,12 @@ const IndexView: React.StatelessComponent<IIndexProps> = ({ title, todos, loadin
         </section>
     </main>;
 
-export default IndexView;
+export default connect<IndexState, IndexDispatch, IndexProps, State>((state): IndexState => ({
+    title: state.index.title,
+    todos: state.index.todos,
+    loading: state.index.loading,
+}), {
+    setTitle,
+    saveTodo,
+    setDone,
+})(Index);
